# datagrid-tester/src/main/resources/api.yaml
openapi: 3.0.0
info:
  description: Datagrid Tester API
  version: "1.0.0"
  title: Datagrid Tester API
  termsOfService: 'http://swagger.io/terms/'
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  '/cache/{cacheName}':
    get:
      tags:
        - cache
      summary: Get cache info
      description: Get cache info
      operationId: getCacheInfo
      parameters:
        - name: cacheName
          in: path
          description: Cache name
          required: true
          schema:
            type: string
            
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CacheInfo'
            
        '400':
          description: Invalid name supplied
        '404':
          description: Cache name not found
...
components:
  schemas:
...
    CacheInfo:
      type: object
      properties:
        name:
          type: string
...
servers:
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/nullp0int3r/DatagridTester/1.0.0